Bug fix that caused the 'exp_mean' values in all datasets to be wrongly transformed and updated the corresponding values.
This is what happend with the data before:

I read the raw, unprocessed dataframe with data scraped from ChEMBL in data_prep.py
I convert the nM values to -log10 by doing: bioactivity = -np.log10(bioactivity) This is fine and causes no issues
I perform some data processing, cliff calculation, splitting, etc. Everything uses the -log10 values (which are fine)
I save all the processed data to a new csv. Here I convert back from -log10 to the original value using: 10 ** abs(np.array(bioactivity)). This is where it goes wrong. For some reason there is an abs() in there that should not be in there. I don't know why this happend, but it did. This is why 10^(abs(-2)) and 10^(abs(2)) both yield 100nM instead of 0.01 and 100. This means that the 'y' column is the correct value and the 'exp_mean' column is wrong.
Luckily, for all model training, evaluation, etc I just use the -log10 values from the 'y' column. This means that the results of the study should stay the same.